+++ archetypes/Exchange/index.tsx
@@ -22,8 +22,8 @@ export default (() => {
     const { swapState, swapDispatch } = useContext(SwapContext);
 
     return (
-        <div className="w-full justify-center mt-14">
-            <div className="bg-theme-background w-full md:w-[611px] md:shadow-xl rounded-3xl py-8 px-4 md:py-8 md:px-12 md:my-8 md:mx-auto ">
+        <div className="w-full justify-center sm:mt-14">
+            <div className="bg-theme-background w-full md:w-[611px] md:shadow-xl sm:rounded-3xl py-8 px-4 md:py-8 md:px-12 md:my-8 md:mx-auto ">
                 <div className="flex">
                     <TWButtonGroup
                         value={swapState?.commitAction ?? CommitActionEnum.mint}
