+++ components/Nav/Navbar/MobileMenu/index.tsx
@@ -75,17 +75,17 @@ export default (({ account, className }) => {
                                             onClick={() => handleRoute('/')}
                                         >
                                             <img className="inline mr-2" src={'/img/general/invest.svg'} alt="Trade" />
-                                            Mint/Burn
+                                            Buy/Sell
                                         </div>
                                         <div
                                             className={classNames(
                                                 linkStyles,
-                                                route === 'balancer-buy-sell' ? selectedStyles : '',
+                                                route === 'mint-burn' ? selectedStyles : '',
                                             )}
-                                            onClick={() => handleRoute('/balancer-buy-sell')}
+                                            onClick={() => handleRoute('/mint-burn')}
                                         >
                                             <img className="inline mr-2" src={'/img/general/invest.svg'} alt="Trade" />
-                                            Buy/Sell
+                                            Mint/Burn
                                         </div>
                                         <div
                                             className={classNames(linkStyles, route === 'browse' ? selectedStyles : '')}
